Add Unset field marker type
This type is used to "mark" that an optional model field is by default
not set but also not required to construct the model. This enables more
semantic usage of built-in dict methods like `get` and `setdefault` that
_can_ return or set a default value. Likewise, `Unset` fields are not
considered to be members of the dictionary
(e.g. `"unset_field" not in model_dict`) and are not included in
`__iter__()`, `keys()`, `values()`, or `len`.

Consider the following:

class Foo(BaseModelDict):
    bar: Optional[int] = None

foo = Foo()

foo.get("bar", 42) # returns None

If you are in the process of refactoring a codebase to use pydantic,
this would be extremely annoying and require extra refactoring work. Now
with the feature:

class Foo(BaseModelDict):
    bar: Optional[int] = Unset

foo = Foo()

assert "bar" not in foo
foo["bar"] # raises KeyError

foo.get("bar", 42) # returns 42
assert "bar" in foo